requirements.txt
accelerate==1.10.1
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
async-timeout==4.0.3
attrs==25.4.0
certifi==2026.2.25
charset-normalizer==3.4.4
click==8.1.8
dataclasses-json==0.6.7
distro==1.9.0
exceptiongroup==1.3.1
faiss-cpu==1.13.0
fastapi==0.128.8
filelock==3.19.1
frozenlist==1.8.0
fsspec==2025.10.0
greenlet==3.2.5
groq==1.0.0
h11==0.16.0
hf-xet==1.3.2
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
httpx-sse==0.4.3
huggingface_hub==0.36.2
idna==3.11
Jinja2==3.1.6
joblib==1.5.3
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.28
langchain-community==0.3.31
langchain-core==0.3.83
langchain-text-splitters==0.3.11
langsmith==0.4.37
MarkupSafe==3.0.3
marshmallow==3.26.2
mpmath==1.3.0
multidict==6.7.1
mypy_extensions==1.1.0
networkx==3.2.1
numpy==1.26.4
ollama==0.6.1
orjson==3.11.5
packaging==25.0
pillow==11.3.0
propcache==0.4.1
psutil==7.2.2
pydantic==2.12.5
pydantic-settings==2.11.0
pydantic_core==2.41.5
PyMuPDF==1.24.1
PyMuPDFb==1.24.1
pypdf==6.8.0
pytesseract==0.3.13
python-dotenv==1.2.1
python-multipart==0.0.9
PyYAML==6.0.3
regex==2026.1.15
requests==2.32.5
requests-toolbelt==1.0.0
safetensors==0.7.0
scikit-learn==1.6.1
scipy==1.13.1
sentence-transformers==5.1.2
sniffio==1.3.1
SQLAlchemy==2.0.48
starlette==0.49.3
sympy==1.14.0
tavily-python==0.7.23
tenacity==9.1.2
threadpoolctl==3.6.0
tiktoken==0.12.0
tokenizers==0.22.2
torch==2.2.2
tqdm==4.67.3
transformers==4.57.6
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uuid_utils==0.14.1
uvicorn==0.39.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==15.0.1
yarl==1.22.0
zstandard==0.25.0
